+++ scripts/generateComponent.js
@@ -2,13 +2,30 @@
 const fs = require('fs');
 
 const COMPONENTS_PATH = './src/components';
+const isDirectory = name => fs.lstatSync(`${COMPONENTS_PATH}/${name}`).isDirectory();
+const currentComponents = fs.readdirSync(COMPONENTS_PATH).map(name => name)
+  .filter(isDirectory);
 
 const componentName = process.argv[process.argv.indexOf('--name') + 1];
+if (currentComponents.indexOf(componentName) > -1) {
+  console.log(`A component with the name ${componentName} already exists!`);
+  return;
+}
+
 const componentPath = `${COMPONENTS_PATH}/${componentName}`;
+const withStyledComponents = process.argv.indexOf('--with-styled-components') > -1;
+const withScss = process.argv.indexOf('--with-scss') > -1;
+let stylesBoilterplate = '// import scss file or styled components';
+
+if (withStyledComponents) {
+  stylesBoilterplate = "import styled from 'styled-components';";
+} else if (withScss) {
+  stylesBoilterplate = `import './${componentName}.scss';`;
+}
 
 const ComponentBoilerplate = `import React from 'react';
 import PropTypes from 'prop-types';
-import './${componentName}.scss';
+${stylesBoilterplate}
 
 const ${componentName} = (props) => (
 
@@ -37,10 +54,12 @@ fs.mkdirSync(componentPath);
 fs.writeFileSync(`${componentPath}/${componentName}.js`, ComponentBoilerplate);
 
 // Create index file.
-fs.writeFileSync(`${componentPath}/index.js`, `export default from './${componentName}';`);
+fs.writeFileSync(`${componentPath}/index.js`, `export default from './${componentName}';\n`);
 
-// Create scss file.
-fs.writeFileSync(`${componentPath}/${componentName}.scss`, ScssBoilerplate);
+if (withScss) {
+  // Create scss file.
+  fs.writeFileSync(`${componentPath}/${componentName}.scss`, ScssBoilerplate);
+}
 
 // Create markdown file.
 fs.writeFileSync(`${componentPath}/${componentName}.md`, '');