+++ src/components/Dropdown/Dropdown.js
@@ -2,14 +2,9 @@ import React, { Component } from 'react';
 import PropTypes from 'prop-types';
 import styled from 'styled-components';
 
+import ToggleMenu from '../ToggleMenu';
 import DropdownItem from '../DropdownItem';
 import Icon from '../Icon';
-
-const Wrapper = styled.div`
-  position: relative;
-  display: inline-block;
-  line-height: 1;
-`;
 
 const DropdownButton = styled.button`
   display: inline-block;
@@ -41,17 +36,13 @@ const ToggleIcon = styled(Icon).attrs({
 `;
 
 const DropdownPanel = styled.div`
-  position: absolute;
-  display: ${(props) => (props.open ? 'block' : 'none')};
   margin-top: 4px;
   background: ${props => props.theme.white};
   border: solid 1px ${props => props.theme.stoneGrey};
   min-width: 100%; // Minimally the width of the dropdown button
-  z-index: 11;
   max-height: 75vh;
   box-shadow: 0 2px 10px ${props => props.theme.stoneGrey};
   overflow: auto;
-  margin-bottom: 1em; // Keep some space at the bottom of the screen
 `;
 
 class Dropdown extends Component {
@@ -59,7 +50,7 @@ class Dropdown extends Component {
     super();
 
     this.state = {
-      showMenu: false,
+      menuOpen: false,
     };
   }
 
@@ -69,47 +60,47 @@ class Dropdown extends Component {
     return isSelected ? 'caret-selected' : 'caret-down';
   };
 
-  showMenu = (event) => {
-    event.preventDefault();
-
-    this.setState({ showMenu: true }, () => {
-      document.addEventListener('click', this.closeMenu);
-    });
+  handleMenuOpen = () => {
+    this.setState({ menuOpen: true });
   };
 
-  closeMenu = (event) => {
-    if (!this.dropdownPanel.contains(event.target)) {
-      this.setState({ showMenu: false }, () => {
-        document.removeEventListener('click', this.closeMenu);
-      });
-    }
+  handleMenuClose = () => {
+    this.setState({ menuOpen: false });
   };
 
-  render() {
-    const { label, className, disabled, isSelected, children } = this.props;
-    const { showMenu } = this.state;
+  renderTrigger = () => {
+    const { disabled, label, isSelected } = this.props;
+    const { menuOpen } = this.state;
 
     return (
-      <Wrapper className={className}>
-        <DropdownButton onClick={this.showMenu} disabled={disabled}>
-          {label}
-          <ToggleIcon
-            open={showMenu}
-            selected={isSelected}
-            name={this.getIcon()}
-            width={isSelected ? 11 : 16}
-            height={9}
-          />
-        </DropdownButton>
-        <DropdownPanel
-          open={this.state.showMenu}
-          innerRef={element => {
-            this.dropdownPanel = element;
-          }}
-        >
+      <DropdownButton disabled={disabled}>
+        {label}
+        <ToggleIcon
+          open={menuOpen}
+          selected={isSelected}
+          name={this.getIcon()}
+          width={isSelected ? 11 : 16}
+          height={9}
+        />
+      </DropdownButton>
+    );
+  }
+
+  render() {
+    const { children, className, position } = this.props;
+
+    return (
+      <ToggleMenu
+        trigger={this.renderTrigger()}
+        className={className}
+        onOpen={this.handleMenuOpen}
+        onClose={this.handleMenuClose}
+        position={position}
+      >
+        <DropdownPanel>
           {children}
         </DropdownPanel>
-      </Wrapper>
+      </ToggleMenu>
     );
   }
 }
@@ -125,6 +116,8 @@ Dropdown.propTypes = {
   isSelected: PropTypes.bool,
   /** Contents of the dropdown */
   children: PropTypes.node.isRequired,
+  /** Position of the dropdown panel */
+  position: PropTypes.oneOf(['top', 'bottom', 'left', 'right'])
 };
 
 Dropdown.defaultProps = {